change project structure
 * move project files into libraries easier to export
