+++ galactuslib/api.go
@@ -8,40 +8,38 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package galactuslib
 
 import (
 	"net/http"
 )
 
-
 // DefaultApiRouter defines the required methods for binding the api requests to a responses for the DefaultApi
-// The DefaultApiRouter implementation should parse necessary information from the http request, 
+// The DefaultApiRouter implementation should parse necessary information from the http request,
 // pass the data to a DefaultApiServicer to perform the required actions, then write the service results to the http response.
-type DefaultApiRouter interface { 
+type DefaultApiRouter interface {
 	GetRoomByCode(http.ResponseWriter, *http.Request)
 }
+
 // RoomApiRouter defines the required methods for binding the api requests to a responses for the RoomApi
-// The RoomApiRouter implementation should parse necessary information from the http request, 
+// The RoomApiRouter implementation should parse necessary information from the http request,
 // pass the data to a RoomApiServicer to perform the required actions, then write the service results to the http response.
-type RoomApiRouter interface { 
+type RoomApiRouter interface {
 	AddRoom(http.ResponseWriter, *http.Request)
 }
 
-
 // DefaultApiServicer defines the api actions for the DefaultApi service
-// This interface intended to stay up to date with the openapi yaml used to generate it, 
-// while the service implementation can ignored with the .openapi-generator-ignore file 
+// This interface intended to stay up to date with the openapi yaml used to generate it,
+// while the service implementation can ignored with the .openapi-generator-ignore file
 // and updated with the logic required for the API.
-type DefaultApiServicer interface { 
+type DefaultApiServicer interface {
 	GetRoomByCode(string) (interface{}, error)
 }
 
-
 // RoomApiServicer defines the api actions for the RoomApi service
-// This interface intended to stay up to date with the openapi yaml used to generate it, 
-// while the service implementation can ignored with the .openapi-generator-ignore file 
+// This interface intended to stay up to date with the openapi yaml used to generate it,
+// while the service implementation can ignored with the .openapi-generator-ignore file
 // and updated with the logic required for the API.
-type RoomApiServicer interface { 
+type RoomApiServicer interface {
 	AddRoom() (interface{}, error)
 }
