+++ galactuslib/api_room_service.go
@@ -8,7 +8,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package galactuslib
 
 import (
 	"errors"
