+++ galactuslib/model_code_and_date.go
@@ -8,10 +8,9 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package galactuslib
 
 type CodeAndDate struct {
-
 	Code string `json:"code,omitempty"`
 
 	Date string `json:"date,omitempty"`
