+++ galactuslib/model_room.go
@@ -8,7 +8,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package galactuslib
 
 import (
 	"time"
@@ -16,7 +16,6 @@ import (
 
 // Room - A room
 type Room struct {
-
 	Id int64 `json:"id,omitempty"`
 
 	Code string `json:"code,omitempty"`
