+++ galactuslib/routers.go
@@ -8,7 +8,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package galactuslib
 
 import (
 	"encoding/json"
@@ -16,6 +16,7 @@ import (
 	"net/http"
 	"os"
 	"strconv"
+
 	"github.com/gorilla/mux"
 )
 
@@ -31,7 +32,7 @@ type Route struct {
 type Routes []Route
 
 // Router defines the required methods for retrieving api routes
-type Router interface { 
+type Router interface {
 	Routes() Routes
 }
