+++ roomapi/api_default.go
@@ -8,13 +8,15 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package roomapi
 
 import (
 	_context "context"
+	"fmt"
 	_ioutil "io/ioutil"
 	_nethttp "net/http"
 	_neturl "net/url"
+	"strings"
 )
 
 // Linger please
@@ -22,17 +24,19 @@ var (
 	_ _context.Context
 )
 
-// RoomApiService RoomApi service
-type RoomApiService service
+// DefaultApiService DefaultApi service
+type DefaultApiService service
 
 /*
-AddRoom Create a new room
+GetRoomByCode Method for GetRoomByCode
+Returns a Room
  * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ * @param code ID of Room to return
 @return Room
 */
-func (a *RoomApiService) AddRoom(ctx _context.Context) (Room, *_nethttp.Response, error) {
+func (a *DefaultApiService) GetRoomByCode(ctx _context.Context, code string) (Room, *_nethttp.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = _nethttp.MethodGet
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -41,7 +45,8 @@ func (a *RoomApiService) AddRoom(ctx _context.Context) (Room, *_nethttp.Response
 	)
 
 	// create path and map variables
-	localVarPath := a.client.cfg.BasePath + "/room"
+	localVarPath := a.client.cfg.BasePath + "/room/{code}"
+	localVarPath = strings.Replace(localVarPath, "{"+"code"+"}", _neturl.QueryEscape(fmt.Sprintf("%v", code)), -1)
 
 	localVarHeaderParams := make(map[string]string)
 	localVarQueryParams := _neturl.Values{}
@@ -93,6 +98,7 @@ func (a *RoomApiService) AddRoom(ctx _context.Context) (Room, *_nethttp.Response
 				return localVarReturnValue, localVarHTTPResponse, newErr
 			}
 			newErr.model = v
+			return localVarReturnValue, localVarHTTPResponse, newErr
 		}
 		return localVarReturnValue, localVarHTTPResponse, newErr
 	}
