+++ roomapi/api_room.go
@@ -8,15 +8,13 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package roomapi
 
 import (
 	_context "context"
 	_ioutil "io/ioutil"
 	_nethttp "net/http"
 	_neturl "net/url"
-	"fmt"
-	"strings"
 )
 
 // Linger please
@@ -24,19 +22,17 @@ var (
 	_ _context.Context
 )
 
-// DefaultApiService DefaultApi service
-type DefaultApiService service
+// RoomApiService RoomApi service
+type RoomApiService service
 
 /*
-GetRoomByCode Method for GetRoomByCode
-Returns a Room
+AddRoom Create a new room
  * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param code ID of Room to return
 @return Room
 */
-func (a *DefaultApiService) GetRoomByCode(ctx _context.Context, code string) (Room, *_nethttp.Response, error) {
+func (a *RoomApiService) AddRoom(ctx _context.Context) (Room, *_nethttp.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = _nethttp.MethodPost
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -45,8 +41,7 @@ func (a *DefaultApiService) GetRoomByCode(ctx _context.Context, code string) (Ro
 	)
 
 	// create path and map variables
-	localVarPath := a.client.cfg.BasePath + "/room/{code}"
-	localVarPath = strings.Replace(localVarPath, "{"+"code"+"}", _neturl.QueryEscape(fmt.Sprintf("%v", code)), -1)
+	localVarPath := a.client.cfg.BasePath + "/room"
 
 	localVarHeaderParams := make(map[string]string)
 	localVarQueryParams := _neturl.Values{}
@@ -98,7 +93,6 @@ func (a *DefaultApiService) GetRoomByCode(ctx _context.Context, code string) (Ro
 				return localVarReturnValue, localVarHTTPResponse, newErr
 			}
 			newErr.model = v
-			return localVarReturnValue, localVarHTTPResponse, newErr
 		}
 		return localVarReturnValue, localVarHTTPResponse, newErr
 	}
