+++ roomapi/configuration.go
@@ -8,7 +8,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package roomapi
 
 import (
 	"net/http"
