+++ roomapi/model_room.go
@@ -8,13 +8,15 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package roomapi
+
 import (
 	"time"
 )
+
 // Room A room
 type Room struct {
-	Id int64 `json:"id,omitempty"`
-	Code string `json:"code,omitempty"`
+	Id        int64     `json:"id,omitempty"`
+	Code      string    `json:"code,omitempty"`
 	CreatedAt time.Time `json:"createdAt,omitempty"`
 }
